// src/types/anime.ts
export type Anime = {
  id: string;
  anilistId: number;
  malId: number | null;
  title: string;
  nativeTitle: string | null;
  romajiTitle: string | null;
  englishTitle: string | null;
  synonyms: string[];
  type: string | null;
  episodes: number | null;
  score: number | null;
  synopsis: string | null;
  imageUrl: string;
  year: number | null;
  bannerImage: string | null;
  status: string | null;
  season: string | null;
  popularity: number | null;
  genres: string[];
};